feat(analyzer-http-client): implement 'delta' watch support (#8)
Allows the client configuration to accept a `watchDeltas` option; when this is set, it will specify the `deltas` query parameter when establishing the SSE connection to an HTTP Server instance.

If the HTTP Server is capable of sending deltas, then the client will listen for them and apply them to the previous state.

This results in a reduced amount of time that the SSE receiver has to spend parsing large events and size of data sent across the wire.