feat: implement configurable retention mechanism
Implements #4

Adds 'retention' options to Analyzer config:

- `maxAge` (number) - the 'age' (in milliseconds) of a Process at which it will be considered 'expired' and deleted from storage.
- `checkInterval` (number) - the time to wait (in milliseconds) between performing expiration runs.

The Analyzer object now contains a `retention` property; this is a Future that will perform a cleanup of Process data older than `maxAge`, every `checkInterval` seconds. This Future represents a long-running process, and as such, it never resolves. However, it will reject if retention has not been explicitly configured via the options above, or if some other error condition is encountered during it's operation. Like all Futures, it must be explicitly forked in order to execute, and may be cancelled to terminate the cancellation process.